Week-06-MLFlow/frontend/requirements.txt
pandas==2.2.3
plotly==5.24.1
gradio==5.7.1
python-dotenv==1.0.1